+++ schemas/sql/b2_trip_stats_by_purpose_employment_age.sql
@@ -12,6 +12,6 @@ CREATE TABLE pt2018.b2_trip_stats_by_purpose_employment_age (
     "purpose_code" INTEGER,
     "purpose" TEXT,
     "trip_count" INTEGER,
-    "trips_per_person" INTEGER,
-    "trips_per_mobile_person" INTEGER
+    "trips_per_person" REAL,
+    "trips_per_mobile_person" REAL
 );
